+++ src/api/announce/error.js
@@ -47,24 +47,10 @@ class BadLocktimeError extends Error {
 	}
 }
 
-/* 
-		Payment Message Errors classes
-*/
-
-class NotValidMessageError extends Error {
-  constructor () {
-    super(`Not valid message.`)
-    this.name = 'NotValidError'
-    this.status = 400
-    this.code = 201 // Error code in hex to avoid confusion with status ?
-  }
-}
-
 module.exports = {
 	MalformedRedeemScriptError,
 	WrongKeyInCLTVError,
 	TheirKeyNotInMultisigError,
-	BadLocktimeError,
-	OurKeyNotInMultisigError,
-	NotValidMessageError
+    OurKeyNotInMultisigError,
+	BadLocktimeError
 }
