+++ src/api/index.js
@@ -1,14 +1,14 @@
-const express = require('express');
+const express = require('express')
 
-const announce = require('./announce');
-const payment = require('./payment');
-const pubkey = require('./pubkey');
+const announce = require('./announce')
+const payment = require('./payment')
+const pubkey = require('./pubkey')
 
-const router = express.Router();
+const router = express.Router()
 
-router.use('/announce', announce);
-router.use('/payment', payment);
-router.use('/pubkey', pubkey);
+router.use('/announce', announce)
+router.use('/payment', payment)
+router.use('/pubkey', pubkey)
 
 
-module.exports = router;
+module.exports = router
